+++ Room.h
@@ -9,7 +9,7 @@ public:
 	GLfloat lowerLeft[2];
 	GLfloat topRight[2];
 	list<Wall*> wallList;
-	Room(GLfloat*, GLfloat*, GLfloat);
+	Room(int*,int*, GLfloat);
 	void addToList(Room*);
 
 private:
